+++ src/telescope.h
@@ -33,7 +33,7 @@ typedef struct torrent {
     uint16_t mtu;
     uint16_t monitorid;
     struct torrent *next;
-    int exclude; // bool
+    uint8_t exclude; // bool
 } torrent_t;
 
 typedef struct telescope_glob {
